api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
boto3==1.35.36
psycopg2-binary==2.9.9
python-dotenv==1.0.1
cachetools==5.5.0
pydantic==2.9.2
